Gracefully handle user cancelation.
Before this commit, when the user canceled the process, the spots
were discarded because an error was generated when the detector could
not find the files corresponding to a timepoint.
Now these missing files are replaced by blanck images, that are
properly handled by the label image detector, and the spots found
before the user canceled can be used in TrackMate later.