Run Cellpose concurrently.
If we have many timepoints and are limited to use Cellpose with CPU
(hello Mac) we could at least try to run the Cellpose detector in a
multithreaded fashion.

This is what we do here: we split all the timepoints to process in
several tasks. Each task first saves the timepoints he got in a folder
specific to the task. Then each task runs its own Python instance and
Cellpose in it. Each Cellpose instance processes the timepoints from
the folder it is points to, and saves the resulting masks in the folder.
In a second step, we sweep all the result folders to rebuild a full result
mask image over all the timepoints.