Try to fetch the value for python path saved in the prefs if we get default.
When we change from the Cellpose detector to another one, the value of
this path, which is precious, is lost and we use the default again.
If we get the default value when we set up the UI, try to fetch the value
from the prefs if any is available.