Fix the parallelization bug.
The reason was that the same instance of the cli was used in
all threads, each setting a different target folder to save the
images. Of course this led to race conditions.

The easy fix was to put the setting of the target folder and
the generation of the command line in a synchronized block.
Not neat but works.